Split order form store
